logging: add logging mechanism, hooks to log uncaught crashes, and an example use of the logging tools (#1)
This adds logging tools in the squid.logging module (along with the
squid package), and two helpers there for getting loggers and setting
the squid logger logging level. This allows us to have a dedicated
hierarchy of loggers that we can configure while leaving all the loggers
setup by other libraries alone.

It also adds an example usage in the control._def.py module. We can
start converting individual files after this.

It also lays the groundwork for better crash reports by allowing us to
hook up handlers for uncaught exceptions. By default these just print to
console, but we want to be able to hook up a FileHandler to our logger
and have the uncaught exceptions go there as well. Actually creating the
FileHandler and attaching it to the root squid logger is to come.

NOTE: This was originally
https://github.com/hongquanli/octopi-research/pull/114 but has been
moved here since this is going to be the main development repository
moving forward.

Tested by:
1. Attempting to run main_hcs.py, and seeing log messages
2. I also added a random raise ValueError("test test") in main_hcs.py
for a different run to make sure that the handlers work as expected
3. And via `python -m pytest`.

---------

Co-authored-by: Ian OHara <ianohara+git@gmail.com>