logging: add a helper for registering a rotating file log that captures all output regardless of stdout log level (#6)
This makes it so we can start logging to a file, and have users send the
file upon crashing. It's setup such that, regardless of
`set_stdout_log_level`, all logs down to `DEBUG` go to the file.

Tested by: Unit tests, and running `main_hcs.py` with and without
`--verbose` and making sure that log output appears in stdout and the
log file as expected.