microcontroller: for simulation, use a stub serial device instead of copy/pasted implementation (#12)
Before, the `Microcontroller_Simulation` implementation was a bunch of
copy/pasted code from the `Microcontroller` simulation. This wasn't
ideal because we needed to keep them in sync for simulation to reflect
reality, and since we needed to have some small differences (eg: no
reader thread) we weren't testing the actual `Microcontroller` code
paths in simulation mode.

this moves the simulation stub down one level to `Serial`, and sets it
up so we can have custom simulated serial implementations in the future
that, for instance, can keep track of positions and act more like the
real firmware.