microcontroller: add the notion of aborted commands, and throw TimeoutError on failed waits (#13)
This changes the command logic so that there's the notion of aborted
commands. These are relevant in 2 cases:
1. Internal to the `Microcontroller` implementation, if the reader
thread doesn't get a command ack in time or the microcontroller responds
with too many checksum errors, a command is aborted.
2. a user of `Microcontroller` can call `abort_current_command` to force
the current command to abort. This is useful if the wait function throws
a `TimeoutError` - the caller can abort the command and move on.

IMPORTANT NOTE: right now, aborting a command does not tell the
microcontroller to do anything. AKA if a motion is in progress, it'll
continue. We could change this behavior but I'm not sure what the
correct behavior is.