gui_hcs: if the microcontroller times out during initialization, reset it before failing (#14)
Before, our gui wait function without an argument would just hang
forever if a command didn't complete. Now, all waits time out at some
point. And if a wait times out during gui initialization, we reset the
micro and re-throw.

We still aren't closing down properly because (I think) the main thread
isn't where this throws. But it is at least really clear in the terminal
that initialization timed out.

This also makes it so other actions that use the wait functions in the
gui can time out. For the most part, this means the gui can continued to
be used (but in-progress actions will likely continue!).