microcontroller: remove is_simulated at Microcontroller level, and instead implement simulated firmware at Serial level (#33)
We added back in an is_simulated flag in the Microcontroller
implementation. This isn't ideal because it meant we had to go and put
`if self.is_simulated:` in all the `Microcontroller` methods and take
custom action in simulation mode. This meant that none of our regular
path code was tested when running in simulation, and that the
`Microcontroller` code had a lot of noise in it to handle the simulation
case.

This moves the simulation handling down to the Serial / simulated
firmware level by handling and responding to `CMD_SET` commands (like
our firmware does) in our simulated serial device.

Tested by: Unit tests